t651 finished with basic functionalities. TODOs are drag and drop of sections

+++ eScreeningDashboard/src/main/webapp/resources/editors/js/app/controllers/sections/sectionsController.js
@@ -1,20 +1,19 @@
 Editors.controller('sectionsController', ['$timeout', '$scope', '$state', 'SurveySectionService', function ($timeout, $scope, $state, SurveySectionService) {
-    $scope.alerts = [];
-    var reloadSSTimeout = undefined;
+    $scope.msgs = [];
 
     $scope.addSuccessMsg = function (reset, reason) {
         addMsg(reset, 'success', reason);
     };
 
     var clearMsgs = function () {
-        $scope.alerts = [];
+        $scope.msgs = [];
     };
 
     var addMsg = function (reset, type, msg) {
         if (reset) {
             clearMsgs();
         }
-        $scope.alerts.push({type: type, msg: msg});
+        $scope.msgs.push({type: type, text: msg});
     }
 
     $scope.addDangerMsg = function (reset, reason) {
@@ -22,15 +21,14 @@ Editors.controller('sectionsController', ['$timeout', '$scope', '$state', 'Surve
     };
 
     $scope.closeAlert = function (index) {
-        $scope.alerts.splice(index, 1);
+        $scope.msgs.splice(index, 1);
     };
 
 
     $scope.addSection = function () {
-        $scope.ssRows.unshift({id: null, name: 'New Survey Section', description: ''});
-        $scope.addSuccessMsg(true, 'New Survey Section is added locally');
+        $scope.ssRows.unshift({id: null, name: 'New Survey Section', description: '', surveys: []});
+        $scope.addSuccessMsg(true, 'Please click on header \'New Survey Section\' and enter the name and description of new survey section');
     };
-
 
     SurveySectionService.readAllSS()
         .then(function (ssRows) {
@@ -42,57 +40,57 @@ Editors.controller('sectionsController', ['$timeout', '$scope', '$state', 'Surve
 
 
     $scope.delete = function (index) {
+
+        // reusable inner function
+        function applyDeleteAction(index) {
+            $scope.ssRows.splice(index, 1);
+            $scope.addSuccessMsg(true, 'Survey Section ' + $scope.ssRows[index].name + ' deleted successfully');
+        }
+
         var section = $scope.ssRows[index];
+
+        // if user is deleting a newly created section (not previously returned from db)
+        if (section.id == null) {
+            applyDeleteAction(index);
+            return;
+        }
+
+        // user is trying to delete data from database
         if (section.surveys !== undefined && section.surveys.length > 0) {
             $scope.addDangerMsg(true, section.name + ' has ' + section.surveys.length + ' surveys and cannot be removed');
         } else {
             SurveySectionService.deleteSS(section)
                 .then(function () {
-                    $scope.addSuccessMsg(true, 'Survey Section ' + section.name + ' deleted successfully');
+                    applyDeleteAction(index);
                 }, function error(reason) {
                     $scope.addDangerMsg(true, reason);
                 });
         }
     }
 
-    $scope.update = function (section) {
+    function freshSS(index, data) {
+        $scope.ssRows[index] = data;
+    }
+
+    $scope.update = function (index) {
+        var section = $scope.ssRows[index];
         SurveySectionService.updateSS(section)
             .then(function (data) {
-                $scope.addSuccessMsg(true, 'Survey Section ' + section.name + ' modified successfully');
+                freshSS(index, data);
+                $scope.addSuccessMsg(true, 'Survey Section ' + data.name + ' modified successfully');
             }, function error(reason) {
                 $scope.addDangerMsg(true, reason);
             });
     }
 
-    $scope.add = function (section) {
+    $scope.add = function (index) {
+        var section = $scope.ssRows[index];
         SurveySectionService.createSS(section)
             .then(function (data) {
-                $scope.addSuccessMsg(true, 'Survey Section ' + section.name + ' added successfully');
+                freshSS(index, data);
+                $scope.addSuccessMsg(true, 'Survey Section ' + data.name + ' added successfully');
             }, function error(reason) {
                 $scope.addDangerMsg(true, reason);
             });
     }
-
-    // on successful Create, Update, and Delete, SurveySectionService sends messages
-    var reloadSS = function () {
-        if (reloadSSTimeout) $timeout.cancel(reloadSSTimeout);
-
-        reloadSSTimeout = $timeout(function () {
-
-            $scope.$apply(function () {
-                SurveySectionService.readAllSS()
-                    .then(function (ssRows) {
-                        $scope.ssRows = ssRows;
-                        //clearMsgs();
-                    }, function error(reason) {
-                        $scope.addDangerMsg(true, reason);
-                    });
-            });
-        }, 500);
-    }
-
-    $scope.$on('ss:created', reloadSS);
-    $scope.$on('ss:deleted', reloadSS);
-    $scope.$on('ss:updated', reloadSS);
-
 }]);
